fix(warn): Silence C4244 for scripts.

+++ Code/Scripts/Toolkit_Objectives.cpp
@@ -361,7 +361,7 @@ DECLARE_SCRIPT(M00_Global_Objective_Controller_RMV, "Set_Type:int, Set_Status:in
 			f_obj = (float)param;
 			f_obj /= 10.0;
 			i_obj = int(f_obj);
-			i_new = (f_obj - i_obj) * 10;
+			i_new = int((f_obj - i_obj) * 10);
 			switch (i_new) {
 			case 1: Commands->Change_Objective_Type(i_obj, OBJECTIVE_TYPE_PRIMARY);
 			case 2: Commands->Change_Objective_Type(i_obj, OBJECTIVE_TYPE_SECONDARY);
@@ -373,7 +373,7 @@ DECLARE_SCRIPT(M00_Global_Objective_Controller_RMV, "Set_Type:int, Set_Status:in
 			f_obj = (float)param;
 			f_obj /= 10.0;
 			i_obj = int(f_obj);
-			i_new = (f_obj - i_obj) * 10;
+			i_new = int((f_obj - i_obj) * 10);
 			switch (i_new) {
 			case 1: Commands->Set_Objective_Status(i_obj, OBJECTIVE_STATUS_PENDING);
 			case 2: Commands->Set_Objective_Status(i_obj, OBJECTIVE_STATUS_ACCOMPLISHED);
